+++ backend/app/repositories/pdf_chunk.py
@@ -21,13 +21,6 @@ class PDFChunkRepository(BaseRepository[PDFChunk]):
             PDFChunk.pdf_id == pdf_id
         ).order_by(PDFChunk.chunk_number).offset(skip).limit(limit).all()
     
-    def get_by_page(self, pdf_id: int, page_number: int) -> List[PDFChunk]:
-        """Get chunks for a specific page."""
-        return self.db.query(PDFChunk).filter(
-            PDFChunk.pdf_id == pdf_id,
-            PDFChunk.page_number == page_number
-        ).order_by(PDFChunk.chunk_number).all()
-    
     def search_content(self, pdf_id: int, search_term: str, skip: int = 0, limit: int = 100) -> List[PDFChunk]:
         """Search chunks by content."""
         return self.db.query(PDFChunk).filter(
@@ -58,50 +51,12 @@ class PDFChunkRepository(BaseRepository[PDFChunk]):
             PDFChunk.content.ilike(f"%{search_term}%")
         ).count()
     
-    def count_by_pdf(self, pdf_id: int) -> int:
-        """Count chunks for a PDF."""
-        return self.db.query(PDFChunk).filter(PDFChunk.pdf_id == pdf_id).count()
-    
-    def get_chunk_by_number(self, pdf_id: int, chunk_number: int) -> Optional[PDFChunk]:
-        """Get specific chunk by number."""
-        return self.db.query(PDFChunk).filter(
-            PDFChunk.pdf_id == pdf_id,
-            PDFChunk.chunk_number == chunk_number
-        ).first()
-    
-    def delete_by_pdf(self, pdf_id: int) -> int:
-        """Delete all chunks for a PDF."""
-        count = self.db.query(PDFChunk).filter(PDFChunk.pdf_id == pdf_id).count()
-        self.db.query(PDFChunk).filter(PDFChunk.pdf_id == pdf_id).delete()
-        self.db.commit()
-        return count
-    
-    def get_content_stats(self, pdf_id: int) -> dict:
-        """Get content statistics for a PDF."""
-        from sqlalchemy import func
-        
-        result = self.db.query(
-            func.count(PDFChunk.id).label('total_chunks'),
-            func.sum(PDFChunk.word_count).label('total_words'),
-            func.sum(PDFChunk.character_count).label('total_characters'),
-            func.avg(PDFChunk.word_count).label('avg_words_per_chunk'),
-            func.max(PDFChunk.page_number).label('max_page')
-        ).filter(PDFChunk.pdf_id == pdf_id).first()
-        
-        return {
-            'total_chunks': result.total_chunks or 0,
-            'total_words': result.total_words or 0,
-            'total_characters': result.total_characters or 0,
-            'avg_words_per_chunk': float(result.avg_words_per_chunk or 0),
-            'max_page': result.max_page or 0
-        }
     
     def bulk_create(self, chunks_data: List[dict]) -> List[PDFChunk]:
         """Create multiple chunks in bulk."""
         chunks = []
         for chunk_data in chunks_data:
             chunk = PDFChunk(**chunk_data)
-            chunk.update_stats()  # Update word and character counts
             chunks.append(chunk)
         
         self.db.add_all(chunks)
