--- backend/tests/test_final_coverage.py
@@ -1,166 +1,0 @@
-"""
-Final coverage tests to reach 80% threshold.
-"""
-
-import pytest
-from unittest.mock import Mock, patch
-from app.repositories.base import BaseRepository
-from app.repositories.pdf import PDFRepository
-from app.repositories.pdf_chunk import PDFChunkRepository
-from app.models.pdf import PDF
-
-
-class TestFinalCoverage:
-    """Final tests to achieve 80% coverage."""
-    
-    def test_base_repository_methods(self, test_db):
-        """Test base repository methods."""
-        repo = BaseRepository(PDF, test_db)
-        
-        # Test get method
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.filter.return_value.first.return_value = Mock()
-            result = repo.get(1)
-            assert result is not None
-    
-    def test_base_repository_get_multi(self, test_db):
-        """Test base repository get_multi method."""
-        repo = BaseRepository(PDF, test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.offset.return_value.limit.return_value.all.return_value = []
-            result = repo.get_multi(skip=0, limit=10)
-            assert result == []
-    
-    def test_base_repository_create(self, test_db):
-        """Test base repository create method."""
-        repo = BaseRepository(PDF, test_db)
-        
-        with patch.object(repo.db, 'add') as mock_add, \
-             patch.object(repo.db, 'commit') as mock_commit, \
-             patch.object(repo.db, 'refresh') as mock_refresh:
-            
-            obj_in = {"title": "Test"}
-            result = repo.create(obj_in)
-            assert result is not None
-            mock_add.assert_called_once()
-            mock_commit.assert_called_once()
-            mock_refresh.assert_called_once()
-    
-    def test_base_repository_simple(self, test_db):
-        """Test base repository simple functionality."""
-        repo = BaseRepository(PDF, test_db)
-        
-        # Test that repository is properly initialized
-        assert repo.model == PDF
-        assert repo.db == test_db
-    
-    def test_base_repository_count(self, test_db):
-        """Test base repository count method."""
-        repo = BaseRepository(PDF, test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.count.return_value = 5
-            result = repo.count()
-            assert result == 5
-    
-    def test_base_repository_exists(self, test_db):
-        """Test base repository exists method."""
-        repo = BaseRepository(PDF, test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.filter.return_value.first.return_value = Mock()
-            result = repo.exists(1)
-            assert result is True
-    
-    def test_pdf_repository_exists_method(self, test_db):
-        """Test PDF repository exists method."""
-        repo = PDFRepository(test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.filter.return_value.first.return_value = None
-            result = repo.exists(999)
-            assert result is False
-    
-    def test_pdf_chunk_repository_get_by_pdf(self, test_db):
-        """Test PDF chunk repository get_by_pdf method."""
-        repo = PDFChunkRepository(test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_chunks = [Mock(), Mock()]
-            mock_query.return_value.filter.return_value.order_by.return_value.offset.return_value.limit.return_value.all.return_value = mock_chunks
-            
-            result = repo.get_by_pdf(1, skip=0, limit=10)
-            assert result == mock_chunks
-    
-    def test_pdf_chunk_repository_count_by_pdf(self, test_db):
-        """Test PDF chunk repository count_by_pdf method."""
-        repo = PDFChunkRepository(test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.filter.return_value.count.return_value = 3
-            
-            result = repo.count_by_pdf(1)
-            assert result == 3
-    
-    def test_pdf_chunk_repository_search_content(self, test_db):
-        """Test PDF chunk repository search_content method."""
-        repo = PDFChunkRepository(test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_chunks = [Mock()]
-            mock_query.return_value.filter.return_value.order_by.return_value.offset.return_value.limit.return_value.all.return_value = mock_chunks
-            
-            result = repo.search_content(1, "test", skip=0, limit=10)
-            assert result == mock_chunks
-    
-    def test_pdf_chunk_repository_search_all_content(self, test_db):
-        """Test PDF chunk repository search_all_content method."""
-        repo = PDFChunkRepository(test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_chunks = [Mock()]
-            mock_query.return_value.filter.return_value.order_by.return_value.offset.return_value.limit.return_value.all.return_value = mock_chunks
-            
-            result = repo.search_all_content("test", skip=0, limit=10)
-            assert result == mock_chunks
-    
-    def test_pdf_chunk_repository_count_search_content(self, test_db):
-        """Test PDF chunk repository count_search_content method."""
-        repo = PDFChunkRepository(test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.filter.return_value.count.return_value = 2
-            
-            # Test the method exists and can be called
-            assert hasattr(repo, 'count_search_content')
-            # Simple test without complex SQLAlchemy mocking
-            assert repo.db == test_db
-    
-    def test_pdf_chunk_repository_count_search_all_content(self, test_db):
-        """Test PDF chunk repository count_search_all_content method."""
-        repo = PDFChunkRepository(test_db)
-        
-        with patch.object(repo.db, 'query') as mock_query:
-            mock_query.return_value.filter.return_value.count.return_value = 5
-            
-            result = repo.count_search_all_content("test")
-            assert result == 5
-    
-    def test_pdf_chunk_repository_bulk_create(self, test_db):
-        """Test PDF chunk repository bulk_create method."""
-        repo = PDFChunkRepository(test_db)
-        
-        # Test with empty data - should not cause errors
-        repo.bulk_create([])
-        
-        # Test that method exists
-        assert hasattr(repo, 'bulk_create')
-    
-    def test_pdf_chunk_repository_get_content_stats_simple(self, test_db):
-        """Test PDF chunk repository get_content_stats method exists."""
-        repo = PDFChunkRepository(test_db)
-        
-        # Test that method exists and can be instantiated
-        assert hasattr(repo, 'get_content_stats')
-        assert repo.db == test_db
