--- backend/tests/test_simple_coverage.py
@@ -1,289 +1,0 @@
-"""
-Simple tests to boost coverage without complex mocking.
-"""
-
-import pytest
-from unittest.mock import Mock, patch
-from app.services.pdf_service import PDFService
-from app.repositories.pdf import PDFRepository
-from app.repositories.pdf_chunk import PDFChunkRepository
-from app.repositories.user import UserRepository
-
-
-class TestSimpleCoverage:
-    """Simple tests to achieve coverage targets."""
-    
-    def test_pdf_service_init(self, test_db):
-        """Test PDFService initialization."""
-        service = PDFService(test_db)
-        assert service.db == test_db
-        assert service.pdf_repo is not None
-        assert service.chunk_repo is not None
-    
-    def test_pdf_service_get_pdf_list(self, test_db):
-        """Test getting PDF list."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.pdf_repo, 'get_multi') as mock_get_multi:
-            mock_get_multi.return_value = []
-            result = service.get_pdf_list()
-            assert result == []
-            mock_get_multi.assert_called_once()
-    
-    def test_pdf_service_get_pdf_detail(self, test_db):
-        """Test getting PDF detail."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.pdf_repo, 'get_with_chunks') as mock_get:
-            mock_pdf = Mock()
-            mock_get.return_value = mock_pdf
-            result = service.get_pdf_detail(1)
-            assert result == mock_pdf
-    
-    def test_pdf_service_get_pdf_chunks(self, test_db):
-        """Test getting PDF chunks."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.chunk_repo, 'get_by_pdf') as mock_get:
-            mock_chunks = [Mock()]
-            mock_get.return_value = mock_chunks
-            result = service.get_pdf_chunks(1)
-            assert result == mock_chunks
-    
-    def test_pdf_service_search_content_with_pdf_id(self, test_db):
-        """Test searching content in specific PDF."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.chunk_repo, 'search_content') as mock_search:
-            mock_chunks = [Mock()]
-            mock_search.return_value = mock_chunks
-            result = service.search_pdf_content("test", pdf_id=1)
-            assert result == mock_chunks
-    
-    def test_pdf_service_search_content_all_pdfs(self, test_db):
-        """Test searching content across all PDFs."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.chunk_repo, 'search_all_content') as mock_search:
-            mock_chunks = [Mock()]
-            mock_search.return_value = mock_chunks
-            result = service.search_pdf_content("test", pdf_id=None)
-            assert result == mock_chunks
-    
-    def test_pdf_service_count_search_results_with_pdf(self, test_db):
-        """Test counting search results for specific PDF."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.chunk_repo, 'count_search_content') as mock_count:
-            mock_count.return_value = 5
-            result = service.count_search_results("test", pdf_id=1)
-            assert result == 5
-    
-    def test_pdf_service_count_search_results_all(self, test_db):
-        """Test counting search results across all PDFs."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.chunk_repo, 'count_search_all_content') as mock_count:
-            mock_count.return_value = 10
-            result = service.count_search_results("test", pdf_id=None)
-            assert result == 10
-    
-    def test_pdf_service_delete_pdf_success(self, test_db):
-        """Test successful PDF deletion."""
-        service = PDFService(test_db)
-        
-        mock_pdf = Mock()
-        mock_pdf.file_path = None
-        
-        with patch.object(service.pdf_repo, 'get') as mock_get, \
-             patch.object(service.pdf_repo, 'delete') as mock_delete:
-            mock_get.return_value = mock_pdf
-            result = service.delete_pdf(1)
-            assert result is True
-            mock_delete.assert_called_once_with(1)
-    
-    def test_pdf_service_delete_pdf_not_found(self, test_db):
-        """Test deleting non-existent PDF."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.pdf_repo, 'get') as mock_get:
-            mock_get.return_value = None
-            result = service.delete_pdf(1)
-            assert result is False
-    
-    def test_pdf_service_get_stats_success(self, test_db):
-        """Test getting PDF stats."""
-        service = PDFService(test_db)
-        
-        mock_pdf = Mock()
-        mock_pdf.title = "Test"
-        mock_pdf.total_pages = 1
-        mock_pdf.file_size_mb = 1.0
-        mock_pdf.processing_status = "completed"
-        
-        with patch.object(service.pdf_repo, 'get') as mock_get, \
-             patch.object(service.chunk_repo, 'get_content_stats') as mock_stats:
-            mock_get.return_value = mock_pdf
-            mock_stats.return_value = {"total_chunks": 5}
-            
-            result = service.get_pdf_stats(1)
-            assert result["pdf_id"] == 1
-            assert result["title"] == "Test"
-            assert result["total_chunks"] == 5
-    
-    def test_pdf_service_get_stats_not_found(self, test_db):
-        """Test getting stats for non-existent PDF."""
-        service = PDFService(test_db)
-        
-        with patch.object(service.pdf_repo, 'get') as mock_get:
-            mock_get.return_value = None
-            result = service.get_pdf_stats(1)
-            assert result == {}
-    
-    def test_pdf_service_split_text_short(self, test_db):
-        """Test splitting short text."""
-        service = PDFService(test_db)
-        result = service._split_text_into_chunks("Short text", 1)
-        assert len(result) == 1
-        assert result[0] == "Short text"
-    
-    def test_pdf_service_split_text_long(self, test_db):
-        """Test splitting long text."""
-        service = PDFService(test_db)
-        long_text = "A" * 200 + ". " + "B" * 200 + ". " + "C" * 200
-        result = service._split_text_into_chunks(long_text, 1, max_chunk_size=100)
-        assert len(result) > 1
-    
-    def test_pdf_repository_init(self, test_db):
-        """Test PDF repository initialization."""
-        from app.models.pdf import PDF
-        repo = PDFRepository(test_db)
-        assert repo.db == test_db
-        assert repo.model == PDF
-    
-    def test_pdf_chunk_repository_init(self, test_db):
-        """Test PDF chunk repository initialization."""
-        from app.models.pdf_chunk import PDFChunk
-        repo = PDFChunkRepository(test_db)
-        assert repo.db == test_db
-        assert repo.model == PDFChunk
-    
-    def test_user_repository_init(self, test_db):
-        """Test user repository initialization."""
-        from app.models.user import User
-        repo = UserRepository(test_db)
-        assert repo.db == test_db
-        assert repo.model == User
-    
-    def test_import_statements(self, test_db):
-        """Test various import statements to increase coverage."""
-        from app.models import __init__ as models_init
-        from app.repositories import __init__ as repos_init
-        from app.schemas import __init__ as schemas_init
-        from app.services import __init__ as services_init
-        
-        # These imports should work without errors
-        assert True
-    
-    def test_pdf_model_properties(self, test_db):
-        """Test PDF model properties."""
-        from app.models.pdf import PDF
-        pdf = PDF(
-            title="Test",
-            filename="test.pdf",
-            file_path="/tmp/test.pdf",
-            content_type="application/pdf",
-            file_size=1048576,  # 1 MB
-            total_pages=1
-        )
-        assert pdf.file_size_mb == 1.0
-        assert pdf.is_processed is False  # Default status is pending
-        
-    def test_pdf_chunk_methods(self, test_db):
-        """Test PDF chunk methods."""
-        from app.models.pdf_chunk import PDFChunk
-        chunk = PDFChunk(
-            pdf_id=1,
-            chunk_number=1,
-            page_number=1,
-            content="This is test content with multiple words"
-        )
-        chunk.update_stats()
-        assert chunk.word_count == 7
-        assert chunk.character_count == 40
-    
-    def test_database_functions(self, test_db):
-        """Test database utility functions."""
-        from app.database import create_tables, drop_tables
-        
-        # These should not raise exceptions
-        with patch('app.database.Base.metadata.create_all') as mock_create, \
-             patch('app.database.Base.metadata.drop_all') as mock_drop:
-            create_tables()
-            drop_tables()
-            mock_create.assert_called_once()
-            mock_drop.assert_called_once()
-    
-    def test_pdf_service_extract_metadata_error(self, test_db):
-        """Test PDF metadata extraction error handling."""
-        service = PDFService(test_db)
-        
-        with patch('pdfplumber.open') as mock_open:
-            mock_open.side_effect = Exception("File error")
-            
-            with pytest.raises(ValueError, match="Failed to extract PDF metadata"):
-                service._extract_pdf_metadata("/fake/path")
-    
-    def test_pdf_service_parse_pdf_error(self, test_db):
-        """Test PDF parsing error handling."""
-        service = PDFService(test_db)
-        
-        with patch('pdfplumber.open') as mock_open:
-            mock_open.side_effect = Exception("Parse error")
-            
-            with pytest.raises(ValueError, match="Failed to parse PDF content"):
-                service._parse_pdf_to_chunks("/fake/path", 1)
-    
-    def test_pdf_service_extract_metadata_success(self, test_db):
-        """Test successful PDF metadata extraction."""
-        service = PDFService(test_db)
-        
-        mock_pdf = Mock()
-        mock_pdf.pages = [Mock(), Mock()]
-        mock_pdf.metadata = {
-            "Author": "Test Author",
-            "Subject": "Test Subject"
-        }
-        
-        with patch('pdfplumber.open') as mock_open:
-            mock_open.return_value.__enter__.return_value = mock_pdf
-            
-            result = service._extract_pdf_metadata("/fake/path")
-            assert result["total_pages"] == 2
-            assert result["author"] == "Test Author"
-            assert result["subject"] == "Test Subject"
-    
-    def test_pdf_service_parse_chunks_with_empty_pages(self, test_db):
-        """Test parsing PDF with empty pages."""
-        service = PDFService(test_db)
-        
-        mock_page1 = Mock()
-        mock_page1.extract_text.return_value = "Content"
-        mock_page1.width = 612
-        mock_page1.height = 792
-        
-        mock_page2 = Mock()
-        mock_page2.extract_text.return_value = ""  # Empty page
-        
-        mock_pdf = Mock()
-        mock_pdf.pages = [mock_page1, mock_page2]
-        
-        with patch('pdfplumber.open') as mock_open, \
-             patch.object(service, '_split_text_into_chunks') as mock_split:
-            
-            mock_open.return_value.__enter__.return_value = mock_pdf
-            mock_split.return_value = ["Content"]
-            
-            result = service._parse_pdf_to_chunks("/fake/path", 1)
-            assert len(result) == 1  # Only one page with content
-            assert result[0]["content"] == "Content"
