+++ backend/tests/unit/__init__.py
@@ -1,0 +1,1 @@
+# Unit tests package
